+++ spring-with-mysql/src/main/java/com/users/demo/dao/CustomerDao.java
@@ -9,8 +9,8 @@ import java.util.List;
 @Repository
 public interface CustomerDao {
     List getAllCustomers();
-    Customer getCustomerById(long id);
-    void deleteCustomerById(long id);
+    Customer getCustomerById(int id);
+    void deleteCustomerById(int id);
     void insertCustomer(Customer customer);
-    void updateCustomerById(long id,Customer customer);
+    void updateCustomerById(int id,Customer customer);
 }
